backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.9.0
pydantic-settings==2.5.0
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.9
httpx==0.27.0
authlib==1.3.1
itsdangerous==2.2.0
python-dotenv==1.0.1
slowapi==0.1.9
ruff==0.6.0
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==5.0.0
